Refactor code to support energy attributes with old odometer fallback
