+++ mytoyota/vehicle.py
@@ -1,9 +1,10 @@
-"""Vehicle class"""
+"""Vehicle representation"""
 import logging
 from typing import Optional
 
-from mytoyota.const import DOORS, HOOD, KEY, LIGHTS, WINDOWS
-from mytoyota.models import Doors, Hood, Key, Lights, Windows
+from mytoyota.location import ParkingLocation
+from mytoyota.status import Energy, Odometer, Sensors
+from mytoyota.utils import format_odometer
 
 _LOGGER: logging.Logger = logging.getLogger(__package__)
 
@@ -24,117 +25,6 @@ class VehicleStatistics:
         return vars(self)
 
 
-class Odometer:
-    """Odometer representation"""
-
-    mileage: int = None
-    unit: str = None
-    fuel: int = None
-
-    def __init__(self, odometer: Optional[list]) -> None:
-
-        _LOGGER.debug("Raw odometer data: %s", str(odometer))
-
-        if odometer is not None:
-
-            odometer_dict = self._format_odometer(odometer)
-
-            _LOGGER.debug("Formatted odometer data: %s", str(odometer_dict))
-
-            if "mileage" in odometer_dict:
-                self.mileage = odometer_dict["mileage"]
-            if "mileage_unit" in odometer_dict:
-                self.unit = odometer_dict["mileage_unit"]
-            if "Fuel" in odometer_dict:
-                self.fuel = odometer_dict["Fuel"]
-
-    def __str__(self) -> str:
-        return str(self.as_dict())
-
-    def as_dict(self) -> dict:
-        """Return odometer as dict."""
-        return vars(self)
-
-    @staticmethod
-    def _format_odometer(raw: list) -> dict:
-        """Formats odometer information from a list to a dict."""
-        instruments: dict = {}
-        for instrument in raw:
-            instruments[instrument["type"]] = instrument["value"]
-            if "unit" in instrument:
-                instruments[instrument["type"] + "_unit"] = instrument["unit"]
-
-        return instruments
-
-
-class ParkingLocation:
-    """ParkingLocation representation"""
-
-    latitude: float = None
-    longitude: float = None
-    timestamp: int = None
-
-    def __init__(self, parking: Optional[dict]) -> None:
-        _LOGGER.debug("Raw parking location data: %s", str(parking))
-
-        if parking is not None:
-            self.latitude = float(parking["lat"])
-            self.longitude = float(parking["lon"])
-            self.timestamp = int(parking["timestamp"])
-
-    def __str__(self) -> str:
-        return str(self.as_dict())
-
-    def as_dict(self) -> dict:
-        """Return parking location as dict."""
-        return vars(self)
-
-
-class Status:
-    """Vehicle status representation"""
-
-    lights: Optional[Lights] = None
-    hood: Optional[Hood] = None
-    doors: Optional[Doors] = None
-    windows: Optional[Windows] = None
-    key: Optional[Key] = None
-
-    overallstatus: str = None
-    last_updated: str = None
-
-    def __init__(self, status: Optional[dict]):
-
-        _LOGGER.debug("Raw status data: %s", str(status))
-
-        if status is not None:
-            self.overallstatus = (
-                status["overallStatus"] if "overallStatus" in status else None
-            )
-
-            self.lights = Lights(status[LIGHTS]) if LIGHTS in status else None
-            self.hood = Hood(status[HOOD]) if HOOD in status else None
-            self.doors = Doors(status[DOORS]) if DOORS in status else None
-            self.windows = Windows(status[WINDOWS]) if WINDOWS in status else None
-            self.key = Key(status[KEY]) if KEY in status else None
-
-            self.last_updated = status["timestamp"] if "timestamp" in status else None
-
-    def __str__(self) -> str:
-        return str(self.as_dict())
-
-    def as_dict(self) -> dict:
-        """Return as dict."""
-        return {
-            "overallstatus": self.overallstatus,
-            "lights": self.lights.as_dict() if self.lights is not None else None,
-            "hood": self.hood.as_dict() if self.hood is not None else None,
-            "doors": self.doors.as_dict() if self.doors is not None else None,
-            "windows": self.windows.as_dict() if self.windows is not None else None,
-            "key": self.key.as_dict() if self.key is not None else None,
-            "last_updated": self.last_updated,
-        }
-
-
 class Vehicle:  # pylint: disable=too-many-instance-attributes
     """Vehicle representation"""
 
@@ -144,9 +34,10 @@ class Vehicle:  # pylint: disable=too-many-instance-attributes
     is_connected: bool = False
     details: Optional[dict] = None
     odometer: Optional[Odometer] = None
+    energy: Optional[Energy] = None
     parking: Optional[ParkingLocation] = None
+    sensors: Optional[Sensors] = None
     statistics: VehicleStatistics = VehicleStatistics()
-    status: Optional[Status] = None
 
     def __init__(  # pylint: disable=too-many-arguments
         self,
@@ -169,35 +60,41 @@ class Vehicle:  # pylint: disable=too-many-instance-attributes
         _LOGGER.debug("Raw vehicle info: %s", str(vehicle_info))
 
         # Vehicle information
-        if "id" in vehicle_info:
-            self.id = vehicle_info["id"]  # pylint: disable=invalid-name
-        if "vin" in vehicle_info:
-            self.vin = vehicle_info["vin"]
-        if "alias" in vehicle_info:
-            self.alias = vehicle_info["alias"]
+        self.id = vehicle_info.get("id", None)  # pylint: disable=invalid-name
+        self.vin = vehicle_info.get("vin", None)
+        self.alias = vehicle_info.get("alias", None)
 
-        # Format vehicle information.
+        # Format vehicle details.
         self.details = self._format_details(vehicle_info)
 
         if self.is_connected:
-            # Extract odometer information.
-            self.odometer = Odometer(odometer) if odometer else Odometer(None)
 
             _LOGGER.debug("Raw status data: %s", str(status))
 
-            # Extract parking information.
-            self.parking = (
-                ParkingLocation(status["event"])
-                if "event" in status
-                else ParkingLocation(None)
+            # Extract fuellevel/Energy capacity information from status.
+            if "energy" in status:
+                _LOGGER.debug("Using energy data: %s", str(status.get("energy")))
+                self.energy = Energy(status.get("energy"))
+            # Use legacy odometer to get fuel level. Older cars still uses this.
+            elif odometer:
+                _LOGGER.debug("Using legacy odometer data: %s", str(odometer))
+                self.energy = Energy(format_odometer(odometer), True)
+                fueltype = self.details.get("fuel", "Unknown")
+                # PATCH: Toyota Aygo reports wrong type.
+                if fueltype == "1.0P":
+                    fueltype = "Petrol"
+                self.energy.type = fueltype
+
+            # Extract mileage and if the car reports in km or mi
+            self.odometer = (
+                Odometer(format_odometer(odometer)) if odometer else Odometer({})
             )
 
-            # Extracts information from status.
-            self.status = (
-                Status(status["protectionState"])
-                if "protectionState" in status
-                else Status(None)
-            )
+            # Extract parking information from status.
+            self.parking = ParkingLocation(status.get("event", {}))
+
+            # Extracts window, door, lock and other information from status.
+            self.sensors = Sensors(status.get("protectionState", {}))
 
     def __str__(self) -> str:
         return str(self.as_dict())
@@ -210,9 +107,10 @@ class Vehicle:  # pylint: disable=too-many-instance-attributes
             "vin": self.vin,
             "details": self.details,
             "status": {
+                "energy": self.energy.as_dict(),
                 "odometer": self.odometer.as_dict(),
                 "parking": self.parking.as_dict(),
-                "vehicle": self.status.as_dict(),
+                "vehicle": self.sensors.as_dict(),
             },
             "servicesEnabled": {
                 "connectedServices": self.is_connected,
